fix(core): Rust 策略核心安全加固——CSPRNG 种子/域密钥分离/注入转义/fail-closed
- shield.rs：指纹种子改用 getrandom OS CSPRNG（原 SystemTime×PID 可预测且前后半区重复）；种子闭包封装不再全局可读（防超级 Cookie）
- per_site_seed.rs：派生改 SHA-256 域密钥分离（多站点种子不可逆推会话种子）；注入不再内嵌会话种子原文、不注册 window 全局属性
- session_state.rs：to_json 改 serde_json（原 format! 手拼零转义，页面标题可注入）
- capability.rs：origin 白名单改前缀精确匹配（原 contains 子串可绕过）+ 空白名单 fail-closed
- oracle.rs：移除 safe_ 前缀降级放行后门；snapshots/reports 有界
- matcher.rs：输入 16K 上限（防平方级无界分配）+ 文档修正 flat URL 用法
- security_policy.rs：文件名截断 UTF-8 字符边界安全（原字节切片可 panic）
- command_bar/space_routing/ext_proxy：注入 JSON 改 serde_json + javascript: 拒绝 + 转义
- action_policy：条件改 token 边界匹配（原 contains 可绕过）
- https_only/query_strip/util/webgl_spoof/ffi：域名终止符、大小写折叠、userinfo/IPv6 提取、Int32Array、hex 长度校验
- 新增 capability/per_site_seed 回归测试

+++ core/rust-policy-core/src/webgl_spoof.rs
@@ -115,7 +115,7 @@ impl WebGLSpoof {
         case MAX_TEXTURE_SIZE:
           return maxTexSize;
         case MAX_VIEWPORT_DIMS:
-          return new Float32Array([maxViewportW, maxViewportH]);
+          return new Int32Array([maxViewportW, maxViewportH]); // 规范要求 Int32——Float32 可被类型检测识破
         case MAX_RENDERBUFFER_SIZE:
           return maxRenderbuf;
         default:
